backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
google-cloud-bigquery==3.18.0
pandas==2.2.1
db-dtypes==1.2.0
python-dotenv==1.0.1
requests==2.31.0
alembic==1.13.1
gspread==6.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-multipart==0.0.20
